requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-snowflake==5.1.1